Move dictionary.py changes to words.py
